Collapse BinanceController into BinanceService
